+++ applications/fishing-map/src/features/timebar/timebar.selectors.ts
@@ -12,7 +12,7 @@ import {
   resolveDataviewEventsResources,
 } from '@globalfishingwatch/dataviews-client'
 import { geoJSONToSegments, Segment } from '@globalfishingwatch/data-transforms'
-import { selectTimebarEvents, selectTimebarGraph } from 'features/app/app.selectors'
+import { selectTimebarGraph, selectVisibleEvents } from 'features/app/app.selectors'
 import { t } from 'features/i18n/i18n'
 import {
   selectActiveTrackDataviews,
@@ -99,8 +99,8 @@ export const selectTracksGraphs = createSelector(
 )
 
 const selectEventsForTracks = createSelector(
-  [selectActiveTrackDataviews, selectResources, selectTimebarEvents],
-  (trackDataviews, resources, timebarEvents) => {
+  [selectActiveTrackDataviews, selectResources, selectVisibleEvents],
+  (trackDataviews, resources, visibleEvents) => {
     const vesselsEvents = trackDataviews.map((dataview) => {
       const { url: tracksUrl } = resolveDataviewDatasetResource(dataview, DatasetTypes.Tracks)
       // const { url: eventsUrl } = resolveDataviewDatasetResource(dataview, DatasetTypes.Events)
@@ -111,15 +111,19 @@ const selectEventsForTracks = createSelector(
         tracksUrl && resources[tracksUrl]?.status === ResourceStatus.Finished
 
       // Waiting for the tracks resource to be resolved to show the events
-      if (!hasEventData || !tracksResourceResolved || timebarEvents === 'none') {
+      if (
+        !hasEventData ||
+        !tracksResourceResolved ||
+        (Array.isArray(visibleEvents) && visibleEvents?.length === 0)
+      ) {
         return []
       }
 
       const eventsResourcesFiltered = eventsResources.filter(({ dataset }) => {
-        if (timebarEvents === 'all') {
+        if (visibleEvents === 'all') {
           return true
         }
-        return dataset.configuration?.type && dataset.configuration?.type === timebarEvents
+        return dataset.configuration?.type && visibleEvents.includes(dataset.configuration?.type)
       })
 
       const data = eventsResourcesFiltered.flatMap(({ url }) => {
@@ -138,6 +142,7 @@ const selectEventsForTracks = createSelector(
 interface RenderedEvent extends ApiEvent {
   color: string
   description: string
+  descriptionGeneric: string
 }
 
 export const selectEventsWithRenderingInfo = createSelector(
@@ -149,6 +154,7 @@ export const selectEventsWithRenderingInfo = createSelector(
           const vesselName = event.vessel.name || event.vessel.id
 
           let description
+          let descriptionGeneric
           switch (event.type) {
             // case 'encounter':
             //   if (event.encounter && event.encounter.vessel.name) {
@@ -163,20 +169,26 @@ export const selectEventsWithRenderingInfo = createSelector(
               } else {
                 description = `${vesselName} ${t('event.portAction')}`
               }
+              descriptionGeneric = `${vesselName} ${t('event.port')}`
               break
             case 'loitering':
               description = `${vesselName} ${t('event.loiteringAction')}`
+              descriptionGeneric = `${vesselName} ${t('event.loitering')}`
               break
             case 'fishing':
               description = `${vesselName} ${t('event.fishingAction')}`
+              descriptionGeneric = `${vesselName} ${t('event.fishing')}`
               break
             default:
               description = 'Unknown event'
+              descriptionGeneric = 'Unknown event'
           }
           const duration = DateTime.fromMillis(event.end as number)
             .diff(DateTime.fromMillis(event.start as number), ['hours', 'minutes'])
             .toObject()
-          description = `${description} for ${duration.hours}hrs ${Math.round(
+
+          // TODO i18n
+          description = `${description} ${duration.hours}hrs ${Math.round(
             duration.minutes as number
           )}mns`
 
@@ -192,6 +204,7 @@ export const selectEventsWithRenderingInfo = createSelector(
             color,
             colorLabels,
             description,
+            descriptionGeneric,
           }
         })
       }