+++ packages/layer-composer/src/generators/tile-cluster/tile-cluster.ts
@@ -56,7 +56,7 @@ class TileClusterGenerator {
         type: 'circle',
         source: config.id,
         'source-layer': 'points',
-        filter: ['>', ['get', 'count'], 1],
+        filter: ['>', ['get', 'count'], config.duplicatedEventsWorkaround ? 2 : 1],
         paint: {
           'circle-radius': [
             'interpolate',
@@ -82,11 +82,13 @@ class TileClusterGenerator {
         type: 'symbol',
         source: config.id,
         'source-layer': 'points',
-        filter: ['>', ['get', 'count'], 1],
+        filter: ['>', ['get', 'count'], config.duplicatedEventsWorkaround ? 2 : 1],
         layout: {
           'text-size': 14,
           'text-offset': [0, 0.13],
-          'text-field': ['get', 'count'],
+          'text-field': config.duplicatedEventsWorkaround
+            ? ['to-string', ['/', ['number', ['get', 'count']], 2]]
+            : ['get', 'count'],
           'text-font': ['Roboto Medium'],
           'text-allow-overlap': true,
         },
@@ -103,7 +105,7 @@ class TileClusterGenerator {
         type: 'circle',
         source: config.id,
         'source-layer': 'points',
-        filter: ['==', ['get', 'count'], 1],
+        filter: ['<=', ['get', 'count'], config.duplicatedEventsWorkaround ? 2 : 1],
         paint: {
           'circle-color': config.color || '#FAE9A0',
           'circle-radius': 5,
@@ -113,6 +115,7 @@ class TileClusterGenerator {
         metadata: {
           interactive: true,
           generatorId: config.id,
+          uniqueFeatureInteraction: config.duplicatedEventsWorkaround ? true : false,
           group: Group.Cluster,
         },
       },