+++ applications/tile-inspector/CHANGELOG.md
@@ -1,5 +1,20 @@
 # tile-inspector
 
+## 0.1.53
+
+### Patch Changes
+
+- Updated dependencies [ea0e1b8a]
+  - @globalfishingwatch/layer-composer@8.5.0
+
+## 0.1.52
+
+### Patch Changes
+
+- Updated dependencies [060b3854]
+- Updated dependencies [ed8af7ae]
+  - @globalfishingwatch/layer-composer@8.4.0
+
 ## 0.1.51
 
 ### Patch Changes